requirements.txt
asgiref==3.3.1
Django==3.1.4
django-contrib-comments==2.0.0
django-markdownx==3.0.1
importlib-metadata==3.3.0
Markdown==3.3.3
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
